Changed UI of Admin, Added Customer Support in Home and Admin Page.

+++ js/AdminOrders.js
@@ -62,6 +62,7 @@ function createTableOrders(){
             let quantity = result[j]["quantity"];
             let date = result[j]["date"];
             let status = result[j]["status"];
+            let customerId = orders[i]["customerId"];
             let unitPrice = result[j]["product"].price;
             let tableData = document.createElement("td");
             if(j == 0){
@@ -96,6 +97,11 @@ function createTableOrders(){
 
                 tableData = document.createElement("td");
                 tableData.innerHTML = date;
+                tableData.setAttribute("rowspan", result.length);
+                tableData.classList.add("table-secondary");
+                tableRow.appendChild(tableData);
+                tableData = document.createElement("td");
+                tableData.innerHTML = customerId;
                 tableData.setAttribute("rowspan", result.length);
                 tableData.classList.add("table-secondary");
                 tableRow.appendChild(tableData);
@@ -148,21 +154,23 @@ function deleteOrdersByorderId(orderid) {
 }
 
 function deleteOrder(){
-    let tableRow = this.parentElement.parentElement;
-    let rowSpan = tableRow.children[0].getAttribute("rowspan");
-    let id = tableRow.children[0].innerHTML;
-    let index = orders.findIndex(order => order.id == id);
-    deleteOrdersByorderId(id);
-    orders.splice(index, 1);
-    updateOriginalOrdersLocalStorage();
-    tableRow.parentElement.removeChild(tableRow);
-    if(rowSpan){
-        for(let i = 0; i < rowSpan - 1; i++){
-            tableRow = tableRow.nextElementSibling;
-            tableRow.parentElement.removeChild(tableRow);
+    if(confirm("Are You sure. Delete this Order!")){
+        let tableRow = this.parentElement.parentElement;
+        let rowSpan = tableRow.children[0].getAttribute("rowspan");
+        let id = tableRow.children[0].innerHTML;
+        let index = orders.findIndex(order => order.id == id);
+        deleteOrdersByorderId(id);
+        orders.splice(index, 1);
+        updateOriginalOrdersLocalStorage();
+        tableRow.parentElement.removeChild(tableRow);
+        if(rowSpan){
+            for(let i = 0; i < rowSpan - 1; i++){
+                tableRow = tableRow.nextElementSibling;
+                tableRow.parentElement.removeChild(tableRow);
+            }
         }
+        createTableOrders();
     }
-    createTableOrders();
 }
 let searchdiv = document.getElementsByClassName("searchbutton")[0];
 searchdiv.children[0].addEventListener("keyup", function(event){
@@ -239,7 +247,7 @@ function filterTableChecked(criteria){
         rowSpan = tableRows[i].children[0].getAttribute("rowspan");
         if(rowSpan){
             let tableData = tableRows[i].children;
-            if(tableData[7].innerHTML == criteria){
+            if(tableData[8].innerHTML == criteria){
                 for(let j = i; j < i + +rowSpan; j++){
                     tableRows[j].classList.remove("d-none");
                 }
@@ -256,7 +264,7 @@ function filterTableUnChecked(criteria){
         rowSpan = tableRows[i].children[0].getAttribute("rowspan");
         if(rowSpan){
             let tableData = tableRows[i].children;
-            if(tableData[7].innerHTML == criteria){
+            if(tableData[8].innerHTML == criteria){
                 for(let j = i; j < i + +rowSpan; j++){
                     tableRows[j].classList.add("d-none");
                 }
